Refactor: Extract SortToggle into shared component
- Created reusable SortToggle component in src/app/components/sort-toggle
- Extracted sort toggle logic and styles from blog and webinar pages
- Updated facet-controls.tsx and search-controls.tsx to use shared component
- Removed duplicate SortToggle implementations and styles
- Maintained all existing functionality and accessibility features

This addresses code review feedback to consolidate duplicate code.


lint issues
