+++ src/app/pages/webinars/search-controls/search-controls.scss
@@ -1,69 +1,7 @@
 @import 'pattern-library/core/pattern-library/headers';
 
-$control-height: 3.6rem;
-$control-radius: 0.6rem;
-
 .webinar-search-controls {
     display: flex;
     justify-content: flex-end;
     padding: $normal-margin 0 2rem;
-}
-
-.webinar-sort-label {
-    color: text-color(normal);
-    font-size: 1.5rem;
-    font-weight: bold;
-    white-space: nowrap;
-}
-
-// --- Sort segmented control ---
-.webinar-sort-wrap {
-    align-items: center;
-    display: flex;
-    gap: 0.8rem;
-}
-
-.webinar-sort {
-    background-color: ui-color(form-bg);
-    border: 0.1rem solid ui-color(secondary);
-    border-radius: $control-radius;
-    display: inline-flex;
-    gap: 0.3rem;
-    padding: 0.3rem;
-
-    button {
-        align-items: center;
-        background-color: transparent;
-        border: 0;
-        border-radius: #{$control-radius - 0.2rem};
-        color: text-color(normal);
-        cursor: pointer;
-        display: inline-flex;
-        font-family: inherit;
-        font-size: 1.5rem;
-        justify-content: center;
-        line-height: 1;
-        min-height: #{$control-height - 0.8rem};
-        padding: 0 1.6rem;
-        transition: background-color 0.15s ease, color 0.15s ease;
-
-        &:hover {
-            color: text-color(link);
-        }
-
-        &[aria-checked='true'] {
-            background-color: text-color(link);
-            color: text-color(white);
-
-            &:hover {
-                color: text-color(white);
-            }
-        }
-    }
-}
-
-// Guarantee a visible keyboard focus indicator.
-.webinar-sort button:focus-visible {
-    outline: 0.2rem solid ui-color(focus-outline);
-    outline-offset: 0.2rem;
 }
